+++ SourceGenerator/Akade.IndexedSet.InternalSourceGenerator.Tests/ReadTests.cs
@@ -54,18 +54,15 @@ public class ReadTests : VerifyBase
     }
 
     [TestMethod]
-    public async Task Generates_boolean_return_wrapped_in_write_lock_for_primary_key_set()
+    public async Task Generates_single_object_wrapped_in_read_lock_for_primary_key_set()
     {
         await this.VerifySourceGen().ForSource<ConcurrentSetSourceGenerator>("""
          public class IndexedSet<TPrimaryKey, TElement>
          {
-             [Akade.IndexedSet.Concurrency.WriteAccess]
-             /// <summary>
-             /// Attempts to remove an item with the given primary key and returns true, if one was found and removed. Otherwise, false.
-             /// </summary>
-             public bool Remove(TPrimaryKey key)
+             [Akade.IndexedSet.Concurrency.ReadAccess]
+             public TElement Single(TPrimaryKey key)
              {
-                 return TryGetSingle(_primaryKeyAccessor, key, out TElement? result, _primaryKeyIndexName) && Remove(result);
+                 return Single(_primaryKeyAccessor, key, _primaryKeyIndexName);
              }
          }
      """);
