+++ SourceGenerator/Akade.IndexedSet.InternalSourceGenerator.Tests/WriteTests.Generates_boolean_return_wrapped_in_write_lock_for_primary_key_set#ConcurrentIndexedSet.write.generated.verified.cs
@@ -1,0 +1,17 @@
+﻿//HintName: ConcurrentIndexedSet.write.generated.cs
+using System.Diagnostics.CodeAnalysis;
+using System.Runtime.CompilerServices;
+
+namespace Akade.IndexedSet.Concurrency;
+#nullable enable
+public partial class ConcurrentIndexedSet<TPrimaryKey, TElement>
+{
+    public bool Remove(TPrimaryKey key)
+    {
+        using (AcquireWriterLock())
+        {
+            return _primaryKeyIndexedSet.Remove(key);
+        }
+    }
+}
+#nullable restore
