+++ SourceGenerator/Akade.IndexedSet.InternalSourceGenerator.Tests/WriteTests.cs
@@ -53,4 +53,22 @@ public class WriteTests : VerifyBase
             }
         """);
     }
+
+    [TestMethod]
+    public async Task Generates_boolean_return_wrapped_in_write_lock_for_primary_key_set()
+    {
+        await this.VerifySourceGen().ForSource<ConcurrentSetSourceGenerator>("""
+     public class IndexedSet<TPrimaryKey, TElement>
+     {
+         [Akade.IndexedSet.Concurrency.WriteAccess]
+         /// <summary>
+         /// Attempts to remove an item with the given primary key and returns true, if one was found and removed. Otherwise, false.
+         /// </summary>
+         public bool Remove(TPrimaryKey key)
+         {
+             return TryGetSingle(_primaryKeyAccessor, key, out TElement? result, _primaryKeyIndexName) && Remove(result);
+         }
+     }
+ """);
+    }
 }
